refactor vote for question
