+++ app/models/answer.rb
@@ -1,6 +1,4 @@
 class Answer < ActiveRecord::Base
-  include Votable
-
   default_scope { order(is_best: :desc, created_at: :asc) }
 
   belongs_to :question
